# Conflicts:
#	DistFiles/Language Explorer/Configuration/Grammar/DataTreeInclude.xml
#	Src/Common/Controls/XMLViews/ReallySimpleListChooser.cs
#	Src/LexText/Interlinear/FocusBoxController.ApproveAndMove.cs
#	Src/LexText/ParserCore/XAmpleParser.cs
#	Src/MasterVersionInfo.txt
#	Src/Transforms/Presentation/FormatXAmpleTrace.xsl
#	Src/xWorks/ConfiguredLcmGenerator.cs
#	Src/xWorks/ILcmContentGenerator.cs
#	Src/xWorks/LcmJsonGenerator.cs
#	Src/xWorks/LcmWordGenerator.cs
#	Src/xWorks/LcmXhtmlGenerator.cs

+++ Lib/src/graphite2/src/inc/GlyphFace.h
@@ -15,8 +15,8 @@
 
     You should also have received a copy of the GNU Lesser General Public
     License along with this library in the file named "LICENSE".
-    If not, write to the Free Software Foundation, 51 Franklin Street, 
-    Suite 500, Boston, MA 02110-1335, USA or visit their web page on the 
+    If not, write to the Free Software Foundation, 51 Franklin Street,
+    Suite 500, Boston, MA 02110-1335, USA or visit their web page on the
     internet at http://www.fsf.org/licenses/lgpl.html.
 
 Alternatively, the contents of this file may be used under the terms of the
@@ -51,7 +51,7 @@ public:
     const Position    & theAdvance() const;
     const Rect        & theBBox() const { return m_bbox; }
     const sparse      & attrs() const { return m_attrs; }
-    uint16              getMetric(uint8 metric) const;
+    int32               getMetric(uint8 metric) const;
 
     CLASS_NEW_DELETE;
 private:
